+++ Tests/BFKitTests/Apple/UIKit/UIViewExtensionTests.swift
@@ -92,6 +92,12 @@ class UIViewExtensionTests: XCTestCase {
         XCTAssertEqual(view.frame, CGRect(x: 0, y: 0, width: 200, height: 100))
     }
     
+    func testSmoothLinearGradientColorsDirection() {
+        view.smoothLinearGradient(colors: [UIColor.red, UIColor.green, UIColor.blue], direction: .custom(startPoint: CGPoint(x: 10, y: 10), endPoint: CGPoint(x: 90, y: 90)))
+        
+        XCTAssertEqual(view.frame, CGRect(x: 0, y: 0, width: 200, height: 100))
+    }
+    
     func testApplyMotionEffect() {
         view.applyMotionEffects()