+++ CHANGELOG.md
@@ -32,6 +32,7 @@ All notable changes to this project will be documented in this file.<br>
 ### Added
 - Added `init?(iso8601:)` to create a Date from an ISO 8601 String in Date extension
 - Added new functions in NSAttributedString extension to handle left, right, center, justified and natural text alignment
+- Added `smoothLinearGradient(colors:, direction:)` function in UIView extension that uses `CGGradient` instead of `CAGradientLayer`
 
 ### Improved
 - Now `dateString(format: = default)` in Date extension has a new `locale` parameter with a `default` value, so no changes are required